Remove AC_FUNC_SETPGRP from "configure"
The AC_FUNC_SETPGRP macro does not work when cross-compiling and causes
the build to fail. According to the documentation of autoconf, this macro
is only used to define "SETPGRP_VOID" or not, but Ularn does not check
this define at all, so calling AC_FUNC_SETPGRP is completely useless.
Additionally, the autoconf documentation also says: "This macro is
obsolescent, as current systems have a setpgrp whose signature conforms
to Posix. New programs need not use this macro."
Thus let's simply remove this macro from the configure script.