Stack size optimization in main()
When compiling for systems that have a very limited stack size,
it's a bad idea to put large arrays on the stack in the main()
function already. So let's get rid of the big buf[BUFSIZ] array
here and initialize optsfile and ckpfile directly instead (like
it is for example already also done for savefilename).